refactor(vault): all functions in its own file
https://coveord.atlassian.net/browse/CDX-935
