+++ packages/cli/src/lib/snapshot/vaultEntriesFunctions/throwSnapshotMissingVaultEntriesError.ts
@@ -1,0 +1,10 @@
+import {SnapshotMissingVaultEntriesError} from '../../errors/snapshotErrors';
+import {VaultTransferFunctionsParam} from './interfaces';
+
+export function throwSnapshotMissingVaultEntriesError({
+  snapshot,
+  cfg,
+  projectPath,
+}: VaultTransferFunctionsParam): never {
+  throw new SnapshotMissingVaultEntriesError(snapshot, cfg, projectPath);
+}
